requirements.txt
streamlit
pandas
datetime
praw
plotly
requests
textblob
wordcloud
matplotlib
python-dotenv